video_analysis_backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
Bottleneck==1.4.2
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
db-dtypes==1.4.0
decorator==5.1.1
Deprecated==1.2.18
dnspython==2.7.0
docstring_parser==0.16
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gcsfs==2024.12.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-documentai==3.1.0
google-cloud-documentai-toolbox==0.14.1a0
google-cloud-language==2.16.0
google-cloud-resource-manager==1.14.0
google-cloud-speech==2.30.0
google-cloud-storage==2.19.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutabledict==4.2.1
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
llvmlite==0.44.0
lxml==5.3.0
MarkupSafe==3.0.2
moviepy==2.1.2
multidict==6.1.0
numba==0.61.0
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pandas-gbq==0.26.1
pikepdf==8.15.1
pillow==10.4.0
proglog==0.1.10
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydata-google-auth==1.9.1
pymongo==4.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
shapely==2.0.6
simple-websocket==1.1.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.6
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3